% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{Essentially a wrapper for the stat_qq and stat_qq_line
functions in ggplot2}
\usage{
qq_plot(
  x,
  color = "blue",
  xlab = expression(Expected ~ ~-log[10](italic(P))),
  ylab = expression(Observed ~ ~-log[10](italic(P))),
  title
)
}
\arguments{
\item{x}{numeric vector or p values}

\item{color}{color of abline}
}
\value{
An QQ plot
}
\description{
Essentially a wrapper for the stat_qq and stat_qq_line
functions in ggplot2
}
